% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Title}
\usage{
resta(a, b)
}
\arguments{
\item{a}{#primer numero}

\item{b}{#segundo numero}
}
\value{
#resultado de la resta
}
\description{
Title
}
\examples{
# se definen ejemplos de como funciona la función
resta(a=4,b=2)
resta(a=10,b=3)

}
